modified prof schema
