+++ package.json
@@ -23,7 +23,7 @@
   "eslintConfig": {
     "extends": "react-app"
   },
-  "proxy": "http://localhost:3001",
+  "proxy": "http://[::1]:3001",
   "browserslist": [
     ">0.2%",
     "not dead",
