+++ routes/api.js
@@ -7,6 +7,10 @@ const Course = Mongo.Course;
 const Prof = Mongo.Prof;
 const Tutor = Mongo.Tutor;
 
+// import from server folder
+const LoadProfile = require("../server/load_profile");
+const Assign = require("../server/assign");
+
 // retrieve a course detail from the db
 router.get('/courses', function(req, res, next){
     Course.findOne({ 'code': req.query.code, 'type': req.query.type} ).then(function(course){
@@ -19,7 +23,7 @@ router.get('/courses', function(req, res, next){
 
 // retrieve a teaching staff detail from the db 
 router.get('/teachers', function(req, res, next){
-    Prof.findOne( {'name': req.query.name} ).then(function(prof){
+    Prof.findOne( {'initial': req.query.name} ).then(function(prof){
         res.json(prof);
         console.log("GET Request for viewing a teaching staff detail");
         console.log("URL Request Params: " + req.query);
@@ -48,79 +52,7 @@ router.post('/courses', function(req, res, next){
 });
 
 // update a course assignee in the db
-router.put('/courses/assign', function(req, res, next){
-    Course.findOneAndUpdate({'code': req.body.code, 'type': req.body.type},
-        {
-            $push: {
-                "schedule.$[i].slots.$[j].scheduledweek":
-                {
-                    'week': req.body.week,
-                    'assignee': req.body.name
-                }
-            }
-      },
-      {
-          arrayFilters:[{'i.group': req.body.group}, {'j.day': req.body.day}]
-      }
-    ).then(function(data){
-
-        Course.aggregate([
-            {
-                $match: {'code': req.body.code, 'type': req.body.type}
-            },
-            {
-                $project: {
-                    schedule: {
-                        $filter:{
-                            input: {
-                                $map: {
-                                    input: "$schedule",
-                                    as: "schedule",
-                                    in: {
-                                        group: "$$schedule.group",
-                                        slots: {
-                                            $filter: {
-                                                input: "$$schedule.slots",
-                                                as: "slot",
-                                                cond: { $eq: ["$$slot.day", req.body.day] }
-                                            }  
-                                        }
-                                    }
-                                 }
-                            },        
-                            as: "schedule",
-                            cond: { $eq: ["$$schedule.group", req.body.group ]}
-                        }
-                    }
-                }
-            }
-        ]).then(function(result){
-
-            console.log("return project: "+JSON.stringify(result));
-            var schedule = result[0].schedule[0];
-            var course = {};
-            course.code = req.body.code;
-            course.type = req.body.type;
-            course.group = req.body.group;
-            course.teachingweek = schedule.slots[0].scheduledweek[0].week;
-            course.day = schedule.slots[0].day;
-            course.time = schedule.slots[0].time;
-            course.venue = schedule.slots[0].venue;
-
-            console.log("course: "+JSON.stringify(course));
- 
-            Prof.findOneAndUpdate({name: req.body.name}, {
-                $push: {courses: course}
-            }).then(function(prof){
-                console.log("updated prof: " + JSON.stringify(prof));
-            });
-        });
-        res.json(data);
-        console.log('PUT Request for teaching assignment');
-        console.log("Request Body: " + req.body);
-        console.log("Response: " + res);
-    }).catch(next);
-});
+router.put('/courses/assign', Assign.assign(req, res, next));
 
 // delete a course from the db 
 router.delete('/courses/:id', function(req, res, next){
