+++ src/addProfile/index.js
@@ -13,8 +13,10 @@ class AddProfile extends Component {
         event.preventDefault();
 
         var requestBody = {};
-        requestBody.name = this.refs.name.value;
+        requestBody.initial = this.refs.name.value;
         requestBody.fullname = this.refs.fullname.value;
+        requestBody.title = this.refs.title.value;
+        requestBody.teachingarea = this.refs.teachingarea.value;
         requestBody.email = this.refs.email.value;
 
         fetch('/api/teachers', {
@@ -44,10 +46,14 @@ class AddProfile extends Component {
             <div id="profile-container">
                 <h1> Add a Teaching Staff Profile </h1>
                 <form className="form-group" id="addProfile" onSubmit={this.handleSubmit}>
-                    <label>Enter a teaching staff alias name </label>
+                    <label>Enter a teaching staff initial </label>
                     <input className="form-control" type="text" ref="name" placeholder="e.g.SP" required />
                     <label>Enter a teaching staff fullname </label>
                     <input className="form-control" type="text" ref="fullname" placeholder="e.g.Shum Ping" required />
+                    <label>Enter his / her title </label>
+                    <input className="form-control" type="text" ref="title" placeholder="e.g.Prof" required />
+                    <label>Enter his / her teaching area </label>
+                    <input className="form-control" type="text" ref="teachingarea" placeholder="e.g.Info-Communications" required />
                     <label>Enter his / her email </label>
                     <input className="form-control" type="text" ref="email" placeholder="e.g.epshum@ntu.edu.sg" required />
                     <input className="form-control" type="submit" value="Add a teaching staff" />
